server/requirements.txt
fastapi
uvicorn
rembg
transformers
pillow
matplotlib
onnxruntime
